services/data-ingestor/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "data-ingestor"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "psycopg2-binary",
    "pydantic",
    "pydantic-settings",
    "jinja2",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "tox",
    "ruff",
    "pyright",
    "pytest",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = "venv"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
testpaths = ["test"]
